Designer: KP-7122 decimal->double

+++ src/Tests/WB.Tests.Unit.Designer/BoundedContexts/Designer/UpdateVariableHandlerTest/when_updating_variable_and_all_parameters_specified.cs
@@ -62,6 +62,6 @@ namespace WB.Tests.Unit.Designer.BoundedContexts.Designer.UpdateVariableHandlerT
         private static Guid responsibleId = Guid.Parse("DDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDD");
         private static string variableName = "name";
         private static string variableExpression = "expression";
-        private static VariableType variableType = VariableType.Decimal;
+        private static VariableType variableType = VariableType.Double;
     }
 }
